--- a/source/timemory/components/trip_count/types.hpp
+++ b/source/timemory/components/trip_count/types.hpp
@@ -38,6 +38,8 @@ TIMEMORY_DECLARE_COMPONENT(trip_count)
 //
 TIMEMORY_SET_COMPONENT_API(component::trip_count, project::timemory, os::agnostic)
 //
+TIMEMORY_DEFINE_CONCRETE_TRAIT(base_has_accum, component::trip_count, false_type)
+//
 TIMEMORY_DEFINE_CONCRETE_TRAIT(report_count, component::trip_count, true_type)
 TIMEMORY_DEFINE_CONCRETE_TRAIT(report_mean, component::trip_count, false_type)
 TIMEMORY_DEFINE_CONCRETE_TRAIT(report_self, component::trip_count, false_type)