- Implemented usage of `std::string_view` when compiled with C++17 support
  - This was applied mostly to component bundlers and the settings class
- This resulted in ~20-40% reduction in overhead
- Replaced some exceptions with `exit(EXIT_FAILURE)`
- Replaced some public inheritance with private inheritance 
- Miscellaneous cleanup of unused code

## Overhead Reduction Summary

Values are average overhead (seconds) for each measurement

| Test | Previous overhead | New Overhead | Reduction | Notes |
| ---- | ------------------ | ---------- | --- | --- |
| blank | 1.051000e-06 | 6.440000e-07 | 38.7% | Label "fibonacci", stack-alloc, 25 unique entries |
| blank-ptr | 1.120000e-06 | 6.910000e-07 | 24.2% | Label "fibonacci", heap-alloc, 25 unique entries |
| basic  | 9.040000e-07 | 6.850000e-07 | 24.2% | Label "fibonacci[N]", stack-alloc, 196417 unique entries |
| basic-ptr | 2.816000e-06 | 2.236000e-06 | 20.6% | Label "fibonacci[N]", heap-alloc, 196417 unique entries |

## Comparison Details

10 iterations for `fibonacci(45)` and instrumenting every fibonacci call about `fibonacci(20)` with various types of macros and creating outputs like: 

```console
| >>> |_run/ex_cxx_overhead.cpp:199 [with timing = mode::blank]          |          1 |          1 | wall       | sec        | 3.488832e+00 | 3.488832e+00 | 3.488832e+00 | 3.488832e+00 | 0.000000e+00 |        0.0 |
| >>>   |_fibonacci                                                      |          1 |          2 | wall       | sec        | 3.488773e+00 | 3.488773e+00 | 3.488773e+00 | 3.488773e+00 | 0.000000e+00 |        0.0 |
| >>>     |_fibonacci                                                    |          2 |          3 | wall       | sec        | 3.488769e+00 | 1.744385e+00 | 2.155430e+00 | 2.155430e+00 | 5.813053e-01 |        0.0 |
| >>>       |_fibonacci                                                  |          4 |          4 | wall       | sec        | 3.488765e+00 | 8.721912e-01 | 1.332463e+00 | 1.332463e+00 | 3.406808e-01 |        0.0 |
| >>>         |_fibonacci                                                |          8 |          5 | wall       | sec        | 3.488755e+00 | 4.360944e-01 | 8.237499e-01 | 8.237499e-01 | 1.959045e-01 |        0.0 |
| >>>           |_fibonacci                                              |         16 |          6 | wall       | sec        | 3.488738e+00 | 2.180461e-01 | 5.090498e-01 | 5.090498e-01 | 1.108064e-01 |        0.0 |
| >>>             |_fibonacci                                            |         32 |          7 | wall       | sec        | 3.488705e+00 | 1.090220e-01 | 3.147378e-01 | 3.147378e-01 | 6.181562e-02 |        0.0 |
...
| >>> |_run/ex_cxx_overhead.cpp:199 [with timing = mode::basic]          |          1 |          1 | wall       | sec        | 3.516781e+00 | 3.516781e+00 | 3.516781e+00 | 3.516781e+00 | 0.000000e+00 |        0.0 |
| >>>   |_fibonacci[45]                                                  |          1 |          2 | wall       | sec        | 3.516766e+00 | 3.516766e+00 | 3.516766e+00 | 3.516766e+00 | 0.000000e+00 |        0.0 |
| >>>     |_fibonacci[44]                                                |          1 |          3 | wall       | sec        | 2.174336e+00 | 2.174336e+00 | 2.174336e+00 | 2.174336e+00 | 0.000000e+00 |        0.0 |
| >>>       |_fibonacci[43]                                              |          1 |          4 | wall       | sec        | 1.343379e+00 | 1.343379e+00 | 1.343379e+00 | 1.343379e+00 | 0.000000e+00 |        0.0 |
| >>>         |_fibonacci[42]                                            |          1 |          5 | wall       | sec        | 8.302545e-01 | 8.302545e-01 | 8.302545e-01 | 8.302545e-01 | 0.000000e+00 |        0.0 |
| >>>           |_fibonacci[41]                                          |          1 |          6 | wall       | sec        | 5.133117e-01 | 5.133117e-01 | 5.133117e-01 | 5.133117e-01 | 0.000000e+00 |        0.0 |
...
| >>>                                           |_fibonacci[25]          |          1 |         22 | wall       | sec        | 2.359160e-04 | 2.359160e-04 | 2.359160e-04 | 2.359160e-04 | 0.000000e+00 |        0.9 |
| >>>                                             |_fibonacci[24]        |          1 |         23 | wall       | sec        | 1.469860e-04 | 1.469860e-04 | 1.469860e-04 | 1.469860e-04 | 0.000000e+00 |        2.5 |
| >>>                                               |_fibonacci[23]      |          1 |         24 | wall       | sec        | 9.034300e-05 | 9.034300e-05 | 9.034300e-05 | 9.034300e-05 | 0.000000e+00 |        2.6 |
| >>>                                                 |_fibonacci[22]    |          1 |         25 | wall       | sec        | 5.610000e-05 | 5.610000e-05 | 5.610000e-05 | 5.610000e-05 | 0.000000e+00 |       40.5 |
| >>>                                                   |_fibonacci[21]  |          1 |         26 | wall       | sec        | 3.338900e-05 | 3.338900e-05 | 3.338900e-05 | 3.338900e-05 | 0.000000e+00 |      100.0 |
| >>>                                                 |_fibonacci[21]    |          1 |         25 | wall       | sec        | 3.186800e-05 | 3.186800e-05 | 3.186800e-05 | 3.186800e-05 | 0.000000e+00 |      100.0 |
| >>>                                               |_fibonacci[22]      |          1 |         24 | wall       | sec        | 5.299300e-05 | 5.299300e-05 | 5.299300e-05 | 5.299300e-05 | 0.000000e+00 |       40.1 |
| >>>                                                 |_fibonacci[21]    |          1 |         25 | wall       | sec        | 3.172800e-05 | 3.172800e-05 | 3.172800e-05 | 3.172800e-05 | 0.000000e+00 |      100.0 |
| >>>                                             |_fibonacci[23]        |          1 |         23 | wall       | sec        | 8.677900e-05 | 8.677900e-05 | 8.677900e-05 | 8.677900e-05 | 0.000000e+00 |        2.6 |
```

```console
$ ./ex_cxx_overhead 45 20 10
```

```cpp
using auto_tuple_t  = tim::auto_tuple_t<wall_clock, user_global_bundle>;

int64_t fibonacci(int64_t n, int64_t cutoff)
{
    if(n > cutoff)
    {
        TIMEMORY_BLANK_MARKER(auto_tuple_t, __FUNCTION__);
        return (n < 2) ? n : (fibonacci(n - 1, cutoff) + fibonacci(n - 2, cutoff));
    }
    return fibonacci(n);
}
```

```cpp
int64_t fibonacci(int64_t n, int64_t cutoff)
{
    if(n > cutoff)
    {
        TIMEMORY_BASIC_MARKER(auto_tuple_t, '[', n, ']');
        return (n < 2) ? n : (fibonacci(n - 1, cutoff) + fibonacci(n - 2, cutoff));
    }
    return fibonacci(n);
}
```

### Previous version

```console
Report from 785668 total measurements and 392884 unique measurements: 
    >>>  run/ex_cxx_overhead.cpp:199 [with timing = mode::none]                                             : 1.683829e+01 sec wall, 1.684000e+01 sec cpu,        0.000 KB peak_rss [laps: 5]

    >>>  run/ex_cxx_overhead.cpp:199 [with timing = mode::blank]                                            : 1.787053e+01 sec wall, 1.787000e+01 sec cpu,        0.000 KB peak_rss [laps: 5]
    >>>  difference vs. 25 unique measurements and 196417 total measurements (mode::blank)                  : 2.064473e-01 sec wall, 2.060000e-01 sec cpu,        0.000 KB peak_rss
    >>>  average overhead of 25 unique measurements and 196417 total measurements (mode::blank)             : 1.051000e-06 sec wall, 1.048000e-06 sec cpu,        0.000 KB peak_rss

    >>>  run/ex_cxx_overhead.cpp:199 [with timing = mode::blank_pointer]                                    : 1.793838e+01 sec wall, 1.793000e+01 sec cpu,        0.000 KB peak_rss [laps: 5]
    >>>  difference vs. 25 unique measurements and 196417 total measurements (mode::blank_pointer)          : 2.200190e-01 sec wall, 2.180000e-01 sec cpu,        0.000 KB peak_rss
    >>>  average overhead of 25 unique measurements and 196417 total measurements (mode::blank_pointer)     : 1.120000e-06 sec wall, 1.109000e-06 sec cpu,        0.000 KB peak_rss

    >>>  run/ex_cxx_overhead.cpp:199 [with timing = mode::basic]                                            : 1.772632e+01 sec wall, 1.772000e+01 sec cpu,    31516.000 KB peak_rss [laps: 5]
    >>>  difference vs. 196417 unique measurements and 196417 total measurements (mode::basic)              : 1.776068e-01 sec wall, 1.760000e-01 sec cpu,     6303.200 KB peak_rss
    >>>  average overhead of 196417 unique measurements and 196417 total measurements (mode::basic)         : 9.040000e-07 sec wall, 8.960000e-07 sec cpu,        0.032 KB peak_rss

    >>>  run/ex_cxx_overhead.cpp:199 [with timing = mode::basic_pointer]                                    : 1.960476e+01 sec wall, 1.961000e+01 sec cpu,    30888.000 KB peak_rss [laps: 5]
    >>>  difference vs. 196417 unique measurements and 196417 total measurements (mode::basic_pointer)      : 5.532936e-01 sec wall, 5.540000e-01 sec cpu,     6177.600 KB peak_rss
    >>>  average overhead of 196417 unique measurements and 196417 total measurements (mode::basic_pointer) : 2.816000e-06 sec wall, 2.820000e-06 sec cpu,        0.031 KB peak_rss
```

### String-view Version

```console
Report from 785668 total measurements and 392884 unique measurements: 
    >>>  run/ex_cxx_overhead.cpp:199 [with timing = mode::none]                                             : 1.691360e+01 sec wall, 1.692000e+01 sec cpu,        0.000 KB peak_rss [laps: 5]

    >>>  run/ex_cxx_overhead.cpp:199 [with timing = mode::blank]                                            : 1.754685e+01 sec wall, 1.757000e+01 sec cpu,        0.000 KB peak_rss [laps: 5]
    >>>  difference vs. 25 unique measurements and 196417 total measurements (mode::blank)                  : 1.266485e-01 sec wall, 1.300000e-01 sec cpu,        0.000 KB peak_rss
    >>>  average overhead of 25 unique measurements and 196417 total measurements (mode::blank)             : 6.440000e-07 sec wall, 6.610000e-07 sec cpu,        0.000 KB peak_rss

    >>>  run/ex_cxx_overhead.cpp:199 [with timing = mode::blank_pointer]                                    : 1.759262e+01 sec wall, 1.758000e+01 sec cpu,        0.000 KB peak_rss [laps: 5]
    >>>  difference vs. 25 unique measurements and 196417 total measurements (mode::blank_pointer)          : 1.358038e-01 sec wall, 1.320000e-01 sec cpu,        0.000 KB peak_rss
    >>>  average overhead of 25 unique measurements and 196417 total measurements (mode::blank_pointer)     : 6.910000e-07 sec wall, 6.720000e-07 sec cpu,        0.000 KB peak_rss

    >>>  run/ex_cxx_overhead.cpp:199 [with timing = mode::basic]                                            : 1.758651e+01 sec wall, 1.757000e+01 sec cpu,    31396.000 KB peak_rss [laps: 5]
    >>>  difference vs. 196417 unique measurements and 196417 total measurements (mode::basic)              : 1.345813e-01 sec wall, 1.300000e-01 sec cpu,     6279.200 KB peak_rss
    >>>  average overhead of 196417 unique measurements and 196417 total measurements (mode::basic)         : 6.850000e-07 sec wall, 6.610000e-07 sec cpu,        0.031 KB peak_rss

    >>>  run/ex_cxx_overhead.cpp:199 [with timing = mode::basic_pointer]                                    : 1.910999e+01 sec wall, 1.911000e+01 sec cpu,    30888.000 KB peak_rss [laps: 5]
    >>>  difference vs. 196417 unique measurements and 196417 total measurements (mode::basic_pointer)      : 4.392776e-01 sec wall, 4.380000e-01 sec cpu,     6177.600 KB peak_rss
    >>>  average overhead of 196417 unique measurements and 196417 total measurements (mode::basic_pointer) : 2.236000e-06 sec wall, 2.229000e-06 sec cpu,        0.031 KB peak_rss
```